Use base class for regular option service
